Create a notification controller
getNotifications function returns a notification object for a specific
user

updateNotifications function inserts a new notification into the
notification object

deleteNotificationsObject is a helper function for removeNotification
It deletes the notification object in the db if all of the notification
arrays are empty

removeNotification function removes either a message or a follow
notification from the notification object. A future update will also
remove a various notification.
removeNotification also calls the deleteNotificationsObject function